backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-auth
pandas
numpy
scikit-learn
joblib
requests